text_fix
- text 글씨체 통합 + 수정
- light들 수정

////////////Todo//////////
1. Axe, Gun 인벤토리에 넣기
2. 착용하는 것 작업해보기
