+++ Assets/Basement/Curtain/source/Curtains_60.5/M_Curtains.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 75bfb9365e435014492f7e9c6c1fc010
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
