+++ Assets/Basement/Curtain/source/Curtains_60.5/M_holder.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5ba0aff22e92f394f8d5eeedaec7f961
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
