+++ Assets/Basement/Curtain/source/Curtains_60.5/M_pipe.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2b72a9b864d8a0468a2574098d3db66
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
