+++ Assets/Basement/Curtain/source/Curtains_60.5/M_tip.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0ad1b72b0c0adf46be13a35c2149509
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
