--- Assets/Scripts/UI/TextTrigger/Acquire/Acquire_Text_Trigger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c8986184a616cbb45a345373b831a1e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
