+++ Assets/Scripts/UI/TextTrigger/Acquire/Cape_Text_Trigger.cs
@@ -3,10 +3,15 @@ using System.Collections.Generic;
 using UnityEngine;
 using TMPro;
 
-public class Cape_Text_Trigger : Acquire_Text_Trigger
+public class Cape_Text_Trigger : MonoBehaviour
 {
     [SerializeField]
     private IngameCtrl ingameCtrl;
+
+    [SerializeField]
+    private TextMeshProUGUI m_Acquire_text;
+
+    private bool b_textTrigger;
 
     public bool b_wear_cape = false;
 
@@ -17,6 +22,7 @@ public class Cape_Text_Trigger : Acquire_Text_Trigger
 
     private void Update()
     {
+
         if (ingameCtrl.m_pressZ && b_textTrigger)
         {
             m_player_cape.SetActive(true);
@@ -32,5 +38,26 @@ public class Cape_Text_Trigger : Acquire_Text_Trigger
             b_wear_cape = false;
         }
  
+    }
+
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.gameObject.CompareTag("Player"))
+        {
+            Debug.Log("player enter");
+            b_textTrigger = true;
+            m_Acquire_text.gameObject.SetActive(true);
+            m_Acquire_text.text = "Press " + "<color=yellow>" + "'Z'" + "</color>" + " to aquire " + "Cape";
+        }
+    }
+
+    private void OnTriggerExit(Collider other)
+    {
+        if (other.gameObject.CompareTag("Player"))
+        {
+            Debug.Log("player out");
+            b_textTrigger = false;
+            m_Acquire_text.gameObject.SetActive(false);
+        }
     }
 }
