+++ Assets/legacy/Assets/AssetArchieve/Prefabs/props/ceiling_light.prefab
@@ -27,7 +27,7 @@ Transform:
   m_GameObject: {fileID: 117088}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 16.4, y: 7.82, z: 20.862}
-  m_LocalScale: {x: 0.01, y: 0.01, z: 0.01}
+  m_LocalScale: {x: 0.055166986, y: 0.01, z: 0.026216995}
   m_Children:
   - {fileID: 429998}
   m_Father: {fileID: 0}
@@ -105,7 +105,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 145134}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -2, y: -45, z: 2}
+  m_LocalPosition: {x: 1.6, y: -59.3, z: 2}
   m_LocalScale: {x: 100, y: 100, z: 100}
   m_Children: []
   m_Father: {fileID: 403928}
@@ -124,7 +124,7 @@ Light:
   m_Shape: 0
   m_Color: {r: 0.89705884, g: 0.97870183, b: 1, a: 1}
   m_Intensity: 0.82
-  m_Range: 13.22
+  m_Range: 18.11
   m_SpotAngle: 30
   m_InnerSpotAngle: 21.80208
   m_CookieSize: 10
@@ -162,7 +162,7 @@ Light:
     serializedVersion: 2
     m_Bits: 4294967295
   m_RenderingLayerMask: 1
-  m_Lightmapping: 4
+  m_Lightmapping: 2
   m_LightShadowCasterMode: 0
   m_AreaSize: {x: 1, y: 1}
   m_BounceIntensity: 0.2
